requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-rest-auth
django-allauth
django-cors-headers
django-silk
django-extensions
drf-spectacular
drf-spectacular-sidecar
gunicorn
django_elasticsearch_dsl
django_elasticsearch_dsl_drf
psycopg2-binary
